+++ src/app/app.module.ts
@@ -2,21 +2,24 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 
 import { AppComponent } from './app.component';
-import { FirstComponent } from './components';
+import { FirstComponent, AboutComponent } from './components';
 
+import { CoreModule } from './core';
 import { ProductsModule } from './products';
 import { CartModule } from './cart';
 
 @NgModule({
   declarations: [
     AppComponent,
-    FirstComponent
+    FirstComponent,
+    AboutComponent
   ],
   imports: [
     BrowserModule,
 
-    ProductsModule,
-    CartModule
+    CoreModule,
+    CartModule,
+    ProductsModule
   ],
   providers: [],
   bootstrap: [AppComponent]