+++ src/app/core/models/settings.model.ts
@@ -1,0 +1,5 @@
+export interface SettingsModel {
+  id: number;
+  login: string;
+  email: string;
+}