+++ src/app/shared/shared.module.ts
@@ -3,8 +3,6 @@ import { CommonModule } from '@angular/common';
 
 import { directives } from './directives';
 
-
-
 @NgModule({
   declarations: [
     ...directives